// src/models/movieSize.model.ts
export enum backdropSize {
    w300 = "w300",
    w780 = "w780",
    w1280 = "w1280",
    original = "original"
}
export enum logoSizes {
    w45 = "w45",
    w92 = "w92",
    w154 = "w154",
    w185 = "w185",
    w300 = "w300",
    w500 = "w500",
    original = "original"
}

export enum posterSizes{
    w92 = "w92",
    w154 = "w154",
    w185 = "w185",
    w342 = "w342",
    w500 = "w500",
    w780 = "w780",
    original = "original"
}

export enum profileSizes{
    w45 = "w45",
    w185 = "w185",
    h632 = "h632",
    original = "original"
}

export enum stillSize{
    w92 = "w92",
    w185 = "w185",
    w300 = "w300",
    original = "original"
}